tools: clean up code, fix entity equals and hashcode

+++ tools/interfaceBuilder/src/main/java/interfacebuilder/projects/ProjectJpaRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProjectJpaRepository extends JpaRepository<ProjectEntity, Integer> {
+public interface ProjectJpaRepository extends JpaRepository<ProjectEntity, Long> {
 	// automatically implements common methods
 }
